Abstract stuff into common.py
